# reborn_rentals_backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API – Proyecto X",
        "description": "Documentación de la API con autenticación JWT (Bearer).",
        "contact": {
            "email": "soporte@tudominio.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor base"
        }
    ],
    "paths": {
...
        "/api/auth/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Listar usuarios (solo admin) o ver perfil propio.",
                "operationId": "19cd421b8eb85e26cf16b0d9540a0002",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Buscar por nombre, email o username",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Tamaño de página (solo admin)",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/UserResource"
                                        },
                                        {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/UserResource"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/user": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Crear un usuario (solo admin).",
                "operationId": "1b109512fb8ca10f46d3c39b7da2ecce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/user/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Ver detalles de usuario (propio o admin).",
                "operationId": "20b2195a90130ead8e75ef1381cbfb52",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Actualizar usuario (propio o admin).",
                "operationId": "ef617beb39335513f38fb35e42a6037c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Eliminar usuario (solo admin).",
                "operationId": "8d2e4e718249eea93750ff2a6fe27da0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "name": {
                        "type": "string",
                        "example": "Fernando Almaraz"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Almaraz",
                        "nullable": true
                    },
                    "second_last_name": {
                        "type": "string",
                        "example": "De la Quintana",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "example": "ferchex@example.com"
                    },
                    "username": {
                        "type": "string",
                        "example": "ferchex"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "+59177777777",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "example": "Cochabamba, Bolivia",
                        "nullable": true
                    },
                    "is_admin": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "UserStoreRequest": {
                "required": [
                    "name",
                    "email",
                    "username",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Fernando Almaraz"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Almaraz"
                    },
                    "second_last_name": {
                        "type": "string",
                        "example": "De la Quintana"
                    },
                    "email": {
                        "type": "string",
                        "example": "ferchex@example.com"
                    },
                    "username": {
                        "type": "string",
                        "example": "ferchex"
                    },
                    "password": {
                        "type": "string",
                        "example": "123456"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "example": "123456"
                    },
                    "is_admin": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "UserUpdateRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Fer Almaraz"
                    },
                    "email": {
                        "type": "string",
                        "example": "ferchex@fusionpax.com"
                    },
                    "username": {
                        "type": "string",
                        "example": "ferchex"
                    },
                    "password": {
                        "type": "string",
                        "example": "654321"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "example": "654321"
                    },
                    "is_admin": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Envia el token como: Bearer {tu_token_jwt}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}